issue18 moved profile functions from configure to profiles
