+++ external/profiles/profiles.go
@@ -1,57 +1,60 @@
 package profiles
 
 import (
-	"crypto/sha256"
-	"encoding/hex"
-	"fmt"
 	"github.com/elasticpath/epcc-cli/config"
 	log "github.com/sirupsen/logrus"
-	"net/url"
+	"gopkg.in/ini.v1"
 	"os"
+	"path/filepath"
 )
 
-func GetProfileName() string {
-	if config.Profile != "" {
-		log.Tracef("Using EPCC_PROFILE value for profile %s", config.Profile)
-		return config.Profile
-	} else {
-		u, err := url.Parse(config.Envs.EPCC_API_BASE_URL)
-		profileName := ""
-		if err != nil {
-			result := newSHA256([]byte(config.Envs.EPCC_CLIENT_ID + ":" + config.Envs.EPCC_API_BASE_URL))
-			profileName = hex.EncodeToString(result)
-		} else {
-			profileName = fmt.Sprintf("%s-%s", u.Host, config.Envs.EPCC_CLIENT_ID)
-		}
+//profile name is set to config.Profile in InitConfig
 
-		log.Tracef("Using auto generated profile name %s", profileName)
-
-		return profileName
-	}
-}
-
-func GetProfileDirectory() string {
-	profileDir := GetProfileDataBaseURL() + GetProfileName()
-
-	log.Tracef("Creating profile directory %s", profileDir)
-	if err := os.MkdirAll(profileDir, 0700); err != nil {
-		panic(fmt.Sprintf("Could not create home directory %v", err))
-	}
-
-	return profileDir
-}
-
-func GetProfileDataBaseURL() string {
-	homeDir, err := os.UserHomeDir()
+func getProfileDirectory() string {
+	home, err := os.UserHomeDir()
 	if err != nil {
-		panic(fmt.Sprintf("Could not get user hoem directory home directory %v", err))
+		log.Errorf("could not get home directory")
+		os.Exit(1)
+	}
+	configDir := home + "/.epcc/profiles_data"
+	configDir = filepath.FromSlash(configDir)
+	//built in check if dir exists
+	if err = os.MkdirAll(configDir, 0700); err != nil {
+		log.Errorf("could not make directory")
 	}
 
-	return homeDir + "/.epcc/profiles_data/"
+	return configDir
 }
 
-// NewSHA256 ...
-func newSHA256(data []byte) []byte {
-	hash := sha256.Sum256(data)
-	return hash[:]
+func GetProfilePath() string {
+	configPath := getProfileDirectory()
+	configPath = filepath.FromSlash(configPath + "/config")
+	if _, err := os.Stat(configPath); err != nil {
+		log.Trace("could not find file at " + configPath)
+		file, err := os.Create(configPath)
+		defer file.Close()
+		if err != nil {
+			log.Errorf("could not create file at " + configPath)
+		}
+		log.Trace("creating config file at " + configPath)
+	}
+
+	return configPath
+}
+
+func GetProfile(name string) *config.Env {
+	result := config.Env{}
+	configPath := GetProfilePath()
+	cfg, err := ini.Load(configPath)
+	if err != nil {
+		log.Errorf("could not load file at " + configPath)
+		os.Exit(1)
+	}
+	if !cfg.HasSection(name) {
+		log.Errorf("could not find profile in file")
+		os.Exit(1)
+	}
+	cfg.Section(name).MapTo(&result)
+	return &result
+
 }
